Split out ProofLine into its own header
Moves ProofLine and ProofLineNumber out of proof_logger.hh and into
proof_line.hh / proof_line-fwd.hh.  This breaks the dependency on the
full proof_logger.hh for code that only needs to store or pass proof
line numbers, reducing unnecessary transitive includes.
